requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   datadog-api-client
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
datadog-api-client==2.33.1
    # via -r requirements.in
google-auth==2.38.0
    # via google-genai
google-genai==1.10.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.3
    # via google-genai
pydantic-core==2.33.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via datadog-api-client
requests==2.32.3
    # via google-genai
rsa==4.9
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
typing-extensions==4.13.1
    # via
    #   anyio
    #   datadog-api-client
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.3.0
    # via
    #   datadog-api-client
    #   requests
websockets==15.0.1
    # via google-genai
